requirements.txt
imap-tools
selenium
bs4
pyperclip
mysql-connector-python
paramiko